requirements.txt
anyio==4.4.0
certifi==2024.8.30
exceptiongroup==1.2.2
filelock==3.15.4
fsspec==2024.6.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
pathlib==1.0.1
pillow==10.4.0
python-telegram-bot==21.5
sniffio==1.3.1
sympy==1.13.2
torch==2.4.0
typing_extensions==4.12.2
